First Release of Randomizer (#19)
Application uses a PRNG abstract base class which is implemented by the XORShift, Linear Congruential Generator, and Mersenne Twister classes. These classes are used by the ProgramRunner class to manage the CLI and then call to the chosen random number generator to create random numbers. CMake, CTest, and CPack are used to build, test, and package the application. Container images are included which allows a user to use the application without installing it locally. A README.md and a man page are also included for documentation.